test: split shutdown safety-floor tests under 300-line cap
Move deferred-recovery shutdown contracts into
safety_floor_shutdown_recovery_test.go so neither new runtime
*_test.go file exceeds the AGENTS.md growth guideline.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
